--- llmkira/sdk/filter/__init__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# @Time    : 2023/9/4 下午10:49
-# @Author  : sudoskys
-# @File    : __init__.py
-# @Software: PyCharm
-
-# 不允许导入此模块
-raise ImportError("This module is not allowed to be imported")
-# 去噪声算法
-from .sublimate import Sublimate
