--- llmkira/sdk/filter/api/keyphrase/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import os
-
-CURRENT_PATH = os.path.abspath(__file__)
-DATA_PATH = os.path.dirname(os.path.dirname(CURRENT_PATH))
