--- llmkira/sdk/filter/api/requirements.txt
@@ -1,5 +1,0 @@
-nltk
-numpy
-jieba
-fasttext-wheel
-scikit-learn
