--- llmkira/sdk/filter/langdetect_fasttext/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-from .detect import detect
-
-__all__ = ["detect"]
