--- llmkira/sdk/filter/langdetect_unicode/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-from .langdetect import LangDetector
-
-__langDetector = LangDetector()
-
-
-def detect(text):
-    return __langDetector.detect(text)
-
-
-__all__ = ['detect']
