+++ llmkira/sdk/utils.py
@@ -9,31 +9,31 @@ import asyncio
 import hashlib
 import tempfile
 from bisect import bisect_left
-from typing import Coroutine, Dict, List, Optional
+from typing import Coroutine, Optional
+from urllib.parse import urlparse
 
 import aiohttp
 import ffmpeg
 import nest_asyncio
 import shortuuid
 from loguru import logger
-from telebot import formatting
-from telebot.formatting import escape_markdown
 
 nest_asyncio.apply()
+
+
+def is_valid_url(url):
+    try:
+        result = urlparse(url)
+        return all([result.scheme, result.netloc])
+    except ValueError:
+        return False
 
 
 def __ensure_event_loop():
     try:
         asyncio.get_event_loop()
-    except Exception:
+    except Exception as e:  # noqa
         asyncio.set_event_loop(asyncio.new_event_loop())
-
-
-def dict2message(data: Dict[str, str]) -> List[str]:
-    _res = []
-    for key, value in data.items():
-        _res.append("".join([formatting.mitalic(key), escape_markdown("="), f"`{escape_markdown(value)}`"]))
-    return _res
 
 
 def sync(coroutine: Coroutine):
@@ -56,7 +56,7 @@ def sha1_encrypt(string):
     sha1加密算法
     """
 
-    sha = hashlib.sha1(string.encode('utf-8'))
+    sha = hashlib.sha1(string.encode("utf-8"))
     encrypts = sha.hexdigest()
     return encrypts[:8]
 
@@ -65,19 +65,20 @@ def generate_uid():
     return shortuuid.uuid()[0:8].upper()
 
 
-async def aiohttp_download_file(url,
-                                session: aiohttp.ClientSession = None,
-                                timeout=None,
-                                size_limit=None,
-                                headers=None,
-                                **kwargs):
+async def aiohttp_download_file(
+    url,
+    session: aiohttp.ClientSession = None,
+    timeout=None,
+    size_limit=None,
+    headers=None,
+    **kwargs,
+):
     if not session:
         session = aiohttp.ClientSession()
     async with session as session:
-        async with session.get(url,
-                               timeout=timeout,
-                               headers=headers, **kwargs
-                               ) as response:
+        async with session.get(
+            url, timeout=timeout, headers=headers, **kwargs
+        ) as response:
             if response.status != 200:
                 raise Exception("无法下载文件")
 
@@ -91,12 +92,13 @@ async def aiohttp_download_file(url,
 
 class Ffmpeg(object):
     @staticmethod
-    def convert(*,
-                input_c: str = "mp3",
-                output_c: str = "ogg",
-                stream_data: bytes = None,
-                quiet=False
-                ) -> Optional[bytes]:
+    def convert(
+        *,
+        input_c: str = "mp3",
+        output_c: str = "ogg",
+        stream_data: bytes = None,
+        quiet=False,
+    ) -> Optional[bytes]:
         """
         使用ffmpeg转换音频格式
         :param input_c: 输入音频格式
@@ -108,8 +110,12 @@ class Ffmpeg(object):
             input_c = "." + input_c
         if not output_c.startswith("."):
             output_c = "." + output_c
-        in_fd, temp_filename = tempfile.mkstemp(suffix=input_c, prefix=None, dir=None, text=False)
-        out_fd, out_temp_filename = tempfile.mkstemp(suffix=output_c, prefix=None, dir=None, text=False)
+        in_fd, temp_filename = tempfile.mkstemp(
+            suffix=input_c, prefix=None, dir=None, text=False
+        )
+        out_fd, out_temp_filename = tempfile.mkstemp(
+            suffix=output_c, prefix=None, dir=None, text=False
+        )
         _bytes = None
         try:
             # 写入文件
@@ -117,15 +123,16 @@ class Ffmpeg(object):
                 f.write(stream_data)
             stream = ffmpeg.input(filename=temp_filename)
             if output_c == ".ogg":
-                stream = ffmpeg.output(stream, filename=out_temp_filename, acodec="libopus")
+                stream = ffmpeg.output(
+                    stream, filename=out_temp_filename, acodec="libopus"
+                )
             else:
                 stream = ffmpeg.output(stream, filename=out_temp_filename)
             stream = ffmpeg.overwrite_output(stream)
-            _ = ffmpeg.run(stream_spec=stream,
-                           quiet=quiet
-                           )
+            _ = ffmpeg.run(stream_spec=stream, quiet=quiet)
             # 读取文件
             import os
+
             _bytes = os.read(out_fd, os.path.getsize(out_temp_filename))
             assert _bytes, "ffmpeg convert failed"
         except Exception as e:
@@ -133,6 +140,7 @@ class Ffmpeg(object):
             raise e
         finally:
             import os
+
             os.close(in_fd)
             os.close(out_fd)
             os.remove(out_temp_filename)
